Began fix on error that with table view not populating.
