--- Movie List/ViewController.swift
@@ -1,30 +1,0 @@
-import UIKit
-
-class ViewController: UIViewController, UITableViewDataSource, UITableViewDelegate {
-    
-    let reuseIdentifier = "cell"
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return Model.shared.numberOfRows()
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: reuseIdentifier, for: indexPath)
-        return cell
-    }
-    
-    
-
-    @IBOutlet weak var tableView: UITableView!
-    
-    @IBOutlet weak var textField: UITextField!
-    
-    @IBAction func add(_ sender: Any) {
-        guard let movie = textField.text, !movie.isEmpty else {return}
-        Model.shared.addMovie(movie)
-        
-        
-    }
-    
-    
-    
-}
